[Refactor] Relocate to `domain.place.application.service`
